software-inventory/requirements.txt
requests
datetime
pyfiglet
python-dotenv